requirements.txt
streamlit
google-generativeai
python-dotenv
openai
pdfplumber
tesseract
pdf2image
requests
pillow